+++ src/elements/Button/DisruptiveButton.jsx
@@ -30,14 +30,11 @@ const DisruptiveButton = React.createClass({
     render() {
         // split 'normal' props from button content
         const {children, ...otherProps} = this.props;
-        const useProgressButton =
-            typeof otherProps.progress !== 'undefined' ||
-            typeof otherProps.progressTopic !== 'undefined';
+        const useProgressButton = typeof otherProps.progress !== 'undefined';
 
         // remove unused propTypes from button
         if (!useProgressButton) {
             delete otherProps.progress;
-            delete otherProps.progressTopic;
         }
 
         if (__DEBUG__ && typeof otherProps.accent !== 'undefined') {