+++ ui-test/tests/Progressbar.jsx
@@ -1,40 +1,15 @@
 import React from 'react';
 
-import rxmq from 'ecc-messagebus';
 import {
     Card,
     CardTitle,
     CardContent,
-    AffirmativeButton,
-    DismissiveButton,
-    DisruptiveButton,
     ProgressButton,
     Progressbar,
 } from '../../index';
 
-const progressTestChannel = rxmq.channel('ui-test.progresstest');
-
 const TestProgressbar = React.createClass({
     progress: 0,
-
-    componentDidMount() {
-        window.setTimeout(this.updateProgress, 5000);
-    },
-
-    updateProgress() {
-        this.progress = this.progress + 5;
-        if (this.progress < 101) {
-            progressTestChannel.subject('number').onNext({
-                progressNumber: this.progress,
-            });
-            progressTestChannel.subject('numberAndDate').onNext({
-                progressNumber: this.progress,
-                lastUpdate: 'Last update 13:37.',
-            });
-            window.setTimeout(this.updateProgress, 1000);
-        }
-    },
-
     render() {
         return (
             <Card stretch={false} shadow={4}>
@@ -56,36 +31,6 @@ const TestProgressbar = React.createClass({
                         progress={0}>
                         Run 2
                     </ProgressButton>
-                    <ProgressButton
-                        progress={66}
-                        tooltip="working"
-                        progressTopic={progressTestChannel.subject('number')}>
-                        Run 3
-                    </ProgressButton>
-                </CardContent>
-                <CardContent>
-                    <AffirmativeButton
-                        progressTopic={progressTestChannel.subject('number')}>
-                        Affirm
-                    </AffirmativeButton>
-                    <DismissiveButton
-                        progress={4}
-                        progressTopic={progressTestChannel.subject('number')}>
-                        Dismiss
-                    </DismissiveButton>
-                    <DisruptiveButton progress={66}>Disrupt</DisruptiveButton>
-                    <DismissiveButton
-                        progressTopic={progressTestChannel.subject('number')}
-                        tooltip="Last update 13:27. Working">
-                        other tooltip
-                    </DismissiveButton>
-                    <DismissiveButton
-                        progressTopic={progressTestChannel.subject(
-                            'numberAndDate'
-                        )}
-                        tooltip="Working">
-                        tooltip with lastUpdate
-                    </DismissiveButton>
                 </CardContent>
                 <Progressbar appearLocal progress={15} />
             </Card>