+++ CHANGELOG.md
@@ -3,9 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
 ## [Unreleased]
-
-TODO: add at least one Added, Changed, Deprecated, Removed, Fixed or Security section
-
+### Removed
+- messagebus
 ## [3.3.1] 2018-12-11
 
 ### Changed