+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ## [Unreleased]
 
-TODO: add at least one Added, Changed, Deprecated, Removed, Fixed or Security section
+### Fixed
+
+- fix specifity of CSS selector for reduced Cards sub elements and improve its layout
 
 ## [3.4.0] 2019-01-16